+++ src/pages/send/[transactionId].tsx
@@ -1,16 +1,11 @@
 import * as Label from '@radix-ui/react-label'
-import { useAtomValue } from 'jotai'
-import { useResetAtom } from 'jotai/utils'
 import dynamic from 'next/dynamic'
-import Image from 'next/image'
-import { useRouter } from 'next/router'
-import { FormEvent, Suspense, useState } from 'react'
-import { useEffect } from 'react'
+import { Suspense } from 'react'
 
 import { styled } from '@/../stitches.config'
-import ConnectWithPhoneDialog from '@/components/ConnectWithPhoneDialog'
-import { Button, Input, PendingButton } from '@/components/primitives'
-import { isLoggedInAtom, stateAtom, userDataAtom } from '@/data/wallet'
+import { Button, FlexRowFixed, Input, Spinner } from '@/components/primitives'
+import SwapForm from '@/components/SwapForm'
+import TransactionDetails from '@/components/TransactionDetails'
 import { getTransactionById, Transaction } from '@/db/transactions'
 
 export async function getServerSideProps({ params: { transactionId } }: { params: { transactionId: string } }) {
@@ -30,18 +25,6 @@ export async function getServerSideProps({ params: { transactionId } }: { params
       transaction,
     },
   }
-}
-
-function SubmitButton({ handleLogin }: { handleLogin: () => void }) {
-  const isLoggedIn = useAtomValue(isLoggedInAtom)
-
-  return isLoggedIn ? (
-    <PendingButton>Send</PendingButton>
-  ) : (
-    <Button as="a" href="#" onClick={handleLogin}>
-      Sign in to send
-    </Button>
-  )
 }
 
 const PageWrapper = styled('div', {
@@ -74,162 +57,53 @@ const InputWrapper = styled('div', {
   color: '$gray12',
 })
 
-const CurrencySymbolWrapper = styled('div', {
-  fontSize: '42px',
-
-  color: '$gray12',
+const InvitePendingMessage = styled('p', {
+  padding: '24px',
+  backgroundColor: '$blue2',
+  color: '$blue10',
+  borderRadius: '24px',
 })
 
-const TransactionDetails = styled('div', {
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '$2',
-  color: '$gray10',
-  padding: '16px',
+const SmallText = styled('span', {
+  fontSize: '$1',
+  color: '$gray9',
 })
 
-const FlexRow = styled('div', {
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'space-between',
-  alignItems: 'flex-end',
-  gap: '$2',
+const PendingText = styled(SmallText, {
+  color: '$blue10',
 })
-
-const DarkText = styled('span', { color: '$gray11' })
 
 function SendTransaction({ transaction }: { transaction: Transaction }) {
-  const userData = useAtomValue(userDataAtom)
-  const refreshState = useResetAtom(stateAtom)
-  const router = useRouter()
-
-  const [isLoginModalOpen, setIsLoginModalOpen] = useState(false)
-
-  const handleSend = async (e: ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault()
-    if (!userData) {
-      throw new Error('Not authenticated, please log in first')
-    }
-    const form = new FormData(e.currentTarget)
-    const response = await fetch(`/api/send`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        fromWallet: userData.publicAddress,
-        fromPhoneNumber: userData.phoneNumber,
-        toPhoneNumber: form.get('toPhoneNumber')?.toString(),
-      }),
-    })
-    if (response.status !== 201) {
-      throw new Error('Error while creating transaction, please try again.')
-    }
-    const data = (await response.json()) as Transaction
-    router.push(`/send/${data.id}`)
+  const handleSend = async () => {
+    // @todo make an actual transaction
   }
-
-  const [inputValue, setInputValue] = useState(1000)
-  const [outputValue, setOutputValue] = useState(1000)
-  const [rate] = useState(0.92)
-  const onInputChange = (e: FormEvent<HTMLInputElement>) => {
-    const newPrice = (e.target.value * rate).toFixed(2)
-    setOutputValue(parseInt(newPrice))
-    setInputValue(e.target.value)
-  }
-
-  useEffect(() => {
-    // Update the document title using the browser API
-    const newPrice = (inputValue * rate).toFixed(2)
-    setOutputValue(parseInt(newPrice))
-  }, [inputValue, rate])
-
-  console.log(transaction)
 
   return (
     <PageWrapper>
-      <small>{JSON.stringify(transaction)}</small>
       <StyledSendForm onSubmit={handleSend}>
-        <InputWrapper>
-          <Label.Root htmlFor="youSendValue">You send</Label.Root>
-          <FlexRow>
-            <CurrencySymbolWrapper>$</CurrencySymbolWrapper>
-            <Input
-              name="youSendValue"
-              placeholder="1000"
-              pattern="[0-9]*"
-              type="number"
-              inputMode="numeric"
-              onChange={onInputChange}
-              value={inputValue}
-            />
-            <FlexRow>
-              <Image src="/USD.png" alt="13" width={20} height={20} priority />
-              USD
-            </FlexRow>
-          </FlexRow>
-        </InputWrapper>
-        <InputWrapper>
-          <Label.Root htmlFor="youSendValue">They receive</Label.Root>
-          <FlexRow>
-            <CurrencySymbolWrapper>€</CurrencySymbolWrapper>
-            <Input
-              name="youReceiveValue"
-              placeholder="€920.26"
-              pattern="[0-9]*"
-              type="number"
-              inputMode="numeric"
-              value={outputValue}
-            />
-            <FlexRow>
-              {' '}
-              <Image src="/EUR.png" alt="13" width={20} height={20} priority />
-              EUR
-            </FlexRow>
-          </FlexRow>
-        </InputWrapper>
+        <SwapForm />
         <InputWrapper>
           <Label.Root htmlFor="toPhoneNumber">To</Label.Root>
-          <Input type="tel" name="toPhoneNumber" placeholder="+1 800 888 8888" value="18455980032" />
-          <h3>INPUT PENDING</h3>
-          <p>We will text you when the recipient joins to complete your transfer! You can safely leave this page.</p>
+          <Input type="tel" name="toPhoneNumber" placeholder="+1 800 888 8888" value="18455980032" disabled />
+          <FlexRowFixed>
+            <Spinner />
+            <PendingText>Waiting for them to join</PendingText>
+          </FlexRowFixed>
+          <InvitePendingMessage>
+            We will text you when the recipient joins to complete your transfer! You can safely leave this page.
+          </InvitePendingMessage>
         </InputWrapper>
-        <TransactionDetails>
-          <FlexRow>
-            <span>Current rate</span>
-            <DarkText>$1 = $0.92</DarkText>
-          </FlexRow>
-          <FlexRow>
-            <span>Total fees</span>
-            <DarkText>$1.23</DarkText>
-          </FlexRow>
-          <FlexRow>
-            <span>Arrival</span>
-            <DarkText>In seconds</DarkText>
-          </FlexRow>
-          <p style={{ fontSize: '12px' }}>
-            Our rate (including any fees) is currently 0.14% better than the European Central Bank (ECB). This
-            comparison rate is typically one of the best available.
-          </p>
-        </TransactionDetails>
-        <Suspense fallback={<Button disabled>...</Button>}>
-          <SubmitButton handleLogin={() => setIsLoginModalOpen(true)} />
-        </Suspense>
-
-        <ConnectWithPhoneDialog
-          isOpen={isLoginModalOpen}
-          setIsOpen={(isOpen) => setIsLoginModalOpen(isOpen)}
-          onConnected={refreshState}
-        />
+        <TransactionDetails />
+        <Button>Send</Button>
       </StyledSendForm>
     </PageWrapper>
   )
 }
 
-function Send() {
+function Send({ transaction }: { transaction: Transaction }) {
   return (
     <Suspense fallback={<p>Loading...</p>}>
-      <SendTransaction />
+      <SendTransaction transaction={transaction} />
     </Suspense>
   )
 }
